Refactor: fix remaining newCubes after levelUp
