+++ src/components/Ice/LoadingPlateLayer/index.js
